Move logic to write pivot table part to PivotTablePartPartWriter. It decreases the clutter and hopefully increases encapsulation, so one writer can optimize how to write its own part without another 4k+ LOC.
